% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add4_or.R
\name{add4_or}
\alias{add4_or}
\title{Computes the add-4 odds ratio.}
\usage{
add4_or(exposure, outcome)
}
\arguments{
\item{exposure}{Logical vector (TRUE/FALSE), exposure variable.}

\item{outcome}{Logical vector (TRUE/FALSE), outcome variable.}
}
\value{
Numeric odds ratio value.
}
\description{
Computes the add-4 odds ratio for two logical vectors.
}
\examples{
exposure <- c(TRUE, FALSE, TRUE, FALSE)
outcome <- c(TRUE, TRUE, FALSE, FALSE)
add4_or(exposure, outcome)
}
