% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_nonempty.R
\name{coalesce_nonempty}
\alias{coalesce_nonempty}
\title{Return the first non-missing, non-empty string.}
\usage{
coalesce_nonempty(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A single character value or NA if none found.
}
\description{
Returns the first element of a character vector that is not NA and not an empty string.
}
\examples{
coalesce_nonempty(c(NA, "", "foo", "bar"))
coalesce_nonempty(c(NA, ""))
}
