% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Fit a Gaussian model (estimate mean and variance).}
\usage{
fit_gaussian(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
List with mean (\code{mu_hat}) and variance (\code{sigma2_hat}).
}
\description{
Computes MLE estimates of mean and variance for a numeric vector.
}
\examples{
fit_gaussian(c(1.2, 3.4, 2.2, 5.0))
}
