% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_zero.R
\name{gt_zero}
\alias{gt_zero}
\title{Truncate values below zero.}
\usage{
gt_zero(x)
}
\arguments{
\item{x}{Numeric value or vector.}
}
\value{
Numeric value or vector with negatives set to zero.
}
\description{
Sets negative values in a numeric vector to zero.
}
\examples{
gt_zero(c(-2, 0, 3, -1))
gt_zero(-5)
}
