% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_binary.R
\name{is_binary}
\alias{is_binary}
\title{Check whether a vector is binary.}
\usage{
is_binary(x, strict = FALSE)
}
\arguments{
\item{x}{Any vector.}

\item{strict}{Logical, if TRUE only allows 0/1 or TRUE/FALSE (default FALSE).}
}
\value{
Logical TRUE if binary, FALSE otherwise.
}
\description{
Checks if a vector contains only two unique values (excluding NA).
}
\examples{
is_binary(c(0, 1, 1, 0))
is_binary(c("a", "b", "a"))
is_binary(c(0, 1, 2))
is_binary(c(TRUE, FALSE, TRUE))
is_binary(c(0, 1, NA))
is_binary(c(0, 2), strict = TRUE)
}
