% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_label.R
\name{majority_label}
\alias{majority_label}
\title{Return the most frequent non-NA value.}
\usage{
majority_label(x, na.rm = TRUE)
}
\arguments{
\item{x}{Any vector.}

\item{na.rm}{Logical, whether to exclude NAs (default TRUE).}
}
\value{
Most frequent value or NA if no non-NA values.
}
\description{
Returns the most frequent value in a vector, optionally excluding NAs.
}
\examples{
majority_label(c(1, 2, 2, 3, NA))
majority_label(c("a", "b", "a", NA), na.rm = FALSE)
}
