% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flat_case.R
\name{make_flat_case}
\alias{make_flat_case}
\title{Convert variable names to flat lowercase with no separators.}
\usage{
make_flat_case(var_names)
}
\arguments{
\item{var_names}{Character vector of variable names.}
}
\value{
Cleaned character vector.
}
\description{
Converts a character vector of variable names to lowercase and removes all non-alphanumeric characters.
}
\examples{
make_flat_case(c("My_Var-Name", "Another.Name", "thirdName"))
}
