% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_colnames.R
\name{print_colnames}
\alias{print_colnames}
\title{Print each column name on a new line.}
\usage{
print_colnames(df)
}
\arguments{
\item{df}{Data frame or tibble.}
}
\value{
Invisibly returns the column names.
}
\description{
Prints each column name of a data frame or tibble on a new line.
}
\examples{
print_colnames(mtcars)
}
